requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiohttp>=3.8.5
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 